Fonctionnel avec bouton

+++ Timer.cpp
@@ -1,7 +1,7 @@
 #include "Timer.h"
 #include <Arduino.h>
 
-Timer StartTimer (unsigned int duration) {
+Timer StartTimer (unsigned long duration) {
     Timer t = {duration, millis()};
 
     return  t;
